backend/requirements.txt
fastapi
uvicorn
python-multipart
supabase
google-genai
python-dotenv
fpdf2
pydantic